Hotfix/#57 admin component folder change usecase rename
