+++ src/app/admin/layout.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import AdminLayoutContainer from "./components/AdminLayoutContainer";
-import AdminNavbar from "./components/Navbar/AdminNavbar";
+import AdminLayoutContainer from "../../components/admin/AdminLayoutContainer";
+import AdminNavbar from "../../components/admin/navbar/AdminNavbar";
 
 const AdminLayout = ({children}:{children:React.ReactNode})=>{
   return(
